Support accessing `_estimator_type` on a class
